+++ src/main.rs
@@ -19,10 +19,9 @@
 //! Show CommonMark documents on TTYs.
 
 extern crate base64;
-extern crate pulldown_cmark;
-extern crate structopt;
 #[macro_use]
-extern crate structopt_derive;
+extern crate clap;
+extern crate pulldown_cmark;
 extern crate syntect;
 extern crate termion;
 
@@ -32,13 +31,13 @@ use std::io::{stdin, stdout};
 use std::fs::File;
 use std::error::Error;
 use std::str::FromStr;
-use structopt::StructOpt;
 use pulldown_cmark::Parser;
 use syntect::parsing::SyntaxSet;
 
 mod tty;
 mod highlighting;
 
+/// Colour options, for the --colour option.
 #[derive(Debug)]
 enum Colour {
     Yes,
@@ -68,39 +67,46 @@ impl FromStr for Colour {
     }
 }
 
-#[derive(StructOpt, Debug)]
-struct Arguments {
-    #[structopt(long = "dump-events", help = "Dump events and exit .")] dump_events: bool,
-    #[structopt(long = "columns", help = "Maximum number of columns.  Defaults to terminal width")]
-    columns: Option<u16>,
-    #[structopt(short = "c", long = "colour", help = "Whether to enable colours.",
-                default_value = "auto")]
-    colour: Colour,
-    #[structopt(help = "Input file.  If absent or - read from standard input")]
-    filename: Option<String>,
-}
-
-fn read_input<T: AsRef<str>>(filename: Option<T>) -> std::io::Result<(PathBuf, String)> {
+/// Read input for `filename`.
+///
+/// If `filename` is `-` read from standard input, otherwise try to open and
+/// read the given file.
+fn read_input<T: AsRef<str>>(filename: T) -> std::io::Result<(PathBuf, String)> {
     let cd = std::env::current_dir()?;
     let mut buffer = String::new();
-    match filename {
-        None => {
-            stdin().read_to_string(&mut buffer)?;
-            Ok((cd, buffer))
-        }
-        Some(ref filename) if filename.as_ref() == "-" => {
-            stdin().read_to_string(&mut buffer)?;
-            Ok((cd, buffer))
-        }
-        Some(filename) => {
-            let mut source = File::open(filename.as_ref())?;
-            source.read_to_string(&mut buffer)?;
-            let base_dir = cd.join(filename.as_ref())
-                .parent()
-                .map(|p| p.to_path_buf())
-                .unwrap_or(cd);
-            Ok((base_dir, buffer))
-        }
+
+    if filename.as_ref() == "-" {
+        stdin().read_to_string(&mut buffer)?;
+        Ok((cd, buffer))
+    } else {
+        let mut source = File::open(filename.as_ref())?;
+        source.read_to_string(&mut buffer)?;
+        let base_dir = cd.join(filename.as_ref())
+            .parent()
+            .map(|p| p.to_path_buf())
+            .unwrap_or(cd);
+        Ok((base_dir, buffer))
+    }
+}
+
+fn process_arguments(args: Arguments) -> Result<(), Box<Error>> {
+    let (base_dir, input) = read_input(args.filename)?;
+    let parser = Parser::new(&input);
+
+    if args.dump_events {
+        tty::dump_events(&mut std::io::stdout(), parser)?;
+        Ok(())
+    } else {
+        let syntax_set = SyntaxSet::load_defaults_newlines();
+        tty::push_tty(
+            &mut std::io::stdout(),
+            args.columns,
+            parser,
+            &base_dir,
+            args.format,
+            syntax_set,
+        )?;
+        Ok(())
     }
 }
 
@@ -156,36 +162,69 @@ fn auto_detect_format(force_colours: bool) -> tty::Format {
     }
 }
 
-fn process_arguments(args: Arguments) -> Result<(), Box<Error>> {
-    let (base_dir, input) = read_input(args.filename)?;
-    let parser = Parser::new(&input);
+/// Represent command line arguments.
+#[derive(Debug)]
+struct Arguments {
+    filename: String,
+    format: tty::Format,
+    columns: u16,
+    dump_events: bool,
+}
 
-    if args.dump_events {
-        tty::dump_events(&mut std::io::stdout(), parser)?;
-        Ok(())
-    } else {
-        let columns = args.columns.unwrap_or_else(terminal_columns);
-        let syntax_set = SyntaxSet::load_defaults_newlines();
-
-        let format = match args.colour {
+impl Arguments {
+    /// Create command line arguments from matches.
+    fn from_matches<'a>(matches: clap::ArgMatches<'a>) -> clap::Result<Self> {
+        let format = match value_t!(matches, "colour", Colour)? {
             Colour::No => tty::Format::NoColours,
             Colour::Yes => auto_detect_format(true),
             Colour::Auto => auto_detect_format(false),
         };
-        tty::push_tty(
-            &mut std::io::stdout(),
+        let filename = value_t!(matches, "filename", String)?;
+        let dump_events = matches.is_present("dump_events");
+        let columns = value_t!(matches, "columns", u16)?;
+
+        Ok(Arguments {
+            filename,
             columns,
-            parser,
-            &base_dir,
+            dump_events,
             format,
-            syntax_set,
-        )?;
-        Ok(())
+        })
     }
 }
 
 fn main() {
-    match process_arguments(Arguments::from_args()) {
+    use clap::*;
+    let columns = terminal_columns().to_string();
+    let app = app_from_crate!()
+        .setting(AppSettings::UnifiedHelpMessage)
+        .arg(
+            Arg::with_name("dump_events")
+                .long("dump-events")
+                .help("Dump Markdown parser events and exit"),
+        )
+        .arg(
+            Arg::with_name("columns")
+                .long("columns")
+                .help("Maximum number of columns to use for output")
+                .default_value(&columns),
+        )
+        .arg(
+            Arg::with_name("colour")
+                .short("c")
+                .long("colour")
+                .help("Whether to enable colours.")
+                .possible_values(&["yes", "no", "auto"])
+                .default_value("auto"),
+        )
+        .arg(
+            Arg::with_name("filename")
+                .help("The file to read.  If - read from standard input instead")
+                .default_value("-"),
+        );
+
+    let matches = app.get_matches();
+    let arguments = Arguments::from_matches(matches).unwrap_or_else(|e| e.exit());
+    match process_arguments(arguments) {
         Ok(_) => std::process::exit(0),
         Err(error) => {
             eprintln!("Error: {}", error);